# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Zhyvo API
  version: 0.1.0
  description: REST API for ephemeral event media rooms. All timestamps are UTC ISO 8601 values.
  license:
    name: Proprietary
    identifier: LicenseRef-Proprietary
servers:
  - url: /api/v1
    description: Same-origin API base path
...
paths:
...
  /media/{mediaID}:
    patch:
      operationId: updateMediaCaption
      tags: [Media]
      summary: Add, edit, or clear a media caption as its uploader or room owner
      parameters:
        - $ref: '#/components/parameters/MediaID'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [caption]
              properties:
                caption: {type: string, maxLength: 300, description: Whitespace is trimmed; an empty string clears the caption}
      responses:
        '200':
          description: Updated caption
          content:
            application/json:
              schema: {$ref: '#/components/schemas/CaptionState'}
        '403': {$ref: '#/components/responses/Forbidden'}
        '404': {$ref: '#/components/responses/NotFound'}
        '422': {$ref: '#/components/responses/ValidationError'}
    delete:
      operationId: deleteMedia
      tags: [Media]
      summary: Delete media as its uploader or room owner
      parameters:
        - $ref: '#/components/parameters/MediaID'
      responses:
        '204': {description: Media deleted}
        '403': {$ref: '#/components/responses/Forbidden'}
        '404': {$ref: '#/components/responses/NotFound'}
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
...
    MediaID:
      name: mediaID
      in: path
      required: true
      schema: {type: string, format: uuid}
...
  responses:
...
    Forbidden:
      description: The identity does not have permission
      content: {application/json: {schema: {$ref: '#/components/schemas/ErrorResponse'}}}
    NotFound:
      description: Resource not found
      content: {application/json: {schema: {$ref: '#/components/schemas/ErrorResponse'}}}
...
    ValidationError:
      description: Input validation failed
      content: {application/json: {schema: {$ref: '#/components/schemas/ErrorResponse'}}}
...
  schemas:
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message, request_id]
          properties:
            code: {type: string, example: ROOM_ACCESS_DENIED}
            message: {type: string}
            request_id: {type: string}
...
    CaptionState:
      type: object
      required: [caption, caption_updated_at]
      properties:
        caption: {type: [string, 'null'], maxLength: 300}
        caption_updated_at: {type: [string, 'null'], format: date-time}
...
security:
  - bearerAuth: []